+++ src/constants.ts
@@ -1,7 +1,7 @@
 export const EXTENSION_NAME = "gistpad";
 export const EXTENSION_ID = `vsls-contrib.gistfs`;
 export const FS_SCHEME = "gist";
-export const PLAYGROUND_JSON_FILE = "playground.json";
+export const PLAYGROUND_FILE = "playground.json";
 export const UNTITLED_SCHEME = "untitled";
 export const URI_PATTERN = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/;
 export const ZERO_WIDTH_SPACE = "‎‎\u200B";